Do not compile event_loop_legacy.c
How the scheme with weak and non-weak implementations is supposed to work is not clear to me.
However, due to the way we link executables - in one big, happy group - the non-weak functions end up overriding the weak ones.
We don't use legacy event loop, so just exclude the file to avoid this mess.